requirements.txt
aiogram==3.15.0
APScheduler==3.10.4
python-dotenv==1.0.1
requests==2.32.3
tzdata==2024.2
urllib3==2.2.3
